Refactor global/colors.js
