+++ src/screens/NationCreateContainer/NationCreateScreen/index.js
@@ -25,7 +25,7 @@ import Images from '../../../global/AssetsImages'
 import Loading from '../../../components/common/Loading';
 import ModalDropdown from 'react-native-modal-dropdown'
 
-import Colors from '../../../global/Colors'
+import Colors from '../../../global/colors'
 import styles from './styles'
 import i18n from '../../../global/i18n';
